+++ src/main/java/org/finance/mybtc/apiManager/impl/btc_e/BTC_EEthImpl.java
@@ -69,7 +69,7 @@ public class BTC_EEthImpl extends ABTC_ECoin {
 	 * apiManager.ESymbol, float)
 	 */
 	@Override
-	public boolean exchange(ESymbol toSymbol, float amount) {
+	public boolean exchange(ESymbol toSymbol, double amount) {
 		boolean result = false;
 		switch (toSymbol) {
 		case BTC:
