+++ src/main/java/org/finance/mybtc/apiManager/impl/okcoin/OkcoinCnyImpl.java
@@ -22,8 +22,8 @@ public class OkcoinCnyImpl extends AOkcoin {
 	 * @see org.finance.mybtc.apiManager.IVirtualCoin#getCoinNum()
 	 */
 	@Override
-	public float getCoinNum() {
-		float coinNum = 0;
+	public double getCoinNum() {
+		double coinNum = 0;
 		OkcoinUserInfo okcoinUserInfo = getOkcoinUserInfo();
 		if (okcoinUserInfo != null && okcoinUserInfo.isResult()) {
 			OkcoinFundsInfo info = okcoinUserInfo.getInfo();
