+++ src/main/java/org/finance/mybtc/change/Ichange.java
@@ -11,7 +11,7 @@ import org.finance.mybtc.change.bean.ChangeInfo;
  */
 public interface Ichange {
 	
-	public float preChange(float totalMoney,float  fromCoinBuyPrice,float toCoinSellPrice,float otherPfChangePrice);
+	public float preChange(double totalMoney,double  fromCoinBuyPrice,double toCoinSellPrice,double otherPfChangePrice);
 	
 	public ChangeInfo execExchange();
